Create practice2.py
In this commit, I've delved into the fascinating world of Python, specifically focusing on two essential topics: "for" loops and the fundamentals of lists. The content in this file covers the intricacies of looping through data with "for" loops and how to manipulate collections using lists.  Happy coding and exploring! 🐍📚🔁